cambio contraseña correo y validaciones

+++ pilatescenter/apps/create_user/views.py
@@ -1,3 +1,4 @@
+
 from django.shortcuts import render, redirect
 from django.http import  HttpResponse
 
@@ -29,7 +30,7 @@ from rest_framework.authentication import TokenAuthentication
 #this is necesary to make the relations between Exercise_det and customuser
 from apps.plan.models import Plan
 from apps.exercise.models import Exercise
-from apps.exercise_det.models import Exercise_det
+from apps.exercise_det.models import Exercise_det, update_resumen
 from apps.lesson_det.models import Lesson_det
 from apps.history_det.models import History_det
 from .models import CustomUser
@@ -105,6 +106,7 @@ def modific_user(request, pk):
 	contexto={
 				'form':form,
 				'exercises_det_list': exercises_det,
+				'user_to_modific':user,
 			 }
 	return render(request,'users/modific_user.html', contexto)
 
@@ -113,7 +115,14 @@ class DeleteUserView(View):
 		Pendiente: cuando un usuario se elimina sacarlo de todas las clases
  	"""
 	def get(self, request, *args, **kwargs):
+
+
 		user = CustomUser.objects.get(pk=self.kwargs['pk'])
+
+		if Lesson_det.objects.filter(id_user_fk= user, saw=False).count() > 0:
+			messages.success(self.request, 'No se puede eliminar un usuario con clases programadas', extra_tags='alert-danger')
+			return redirect('content_user:list_user')
+
 		user.delete()
 		return redirect('content_user:list_user')
 
@@ -152,6 +161,11 @@ class LockUserView(View):
  	"""
 	def get(self, request, *args, **kwargs):
 		user = CustomUser.objects.get(pk=self.kwargs['pk'])
+
+		if Lesson_det.objects.filter(id_user_fk= user, saw=False).count() > 0:
+			messages.success(self.request, 'No se puede bloquear un usuario con clases programadas', extra_tags='alert-danger')
+			return redirect('content_user:list_user')
+
 		user.is_active=False
 		user.save()
 		return redirect('content_user:list_user')
@@ -192,8 +206,8 @@ class ResetUsersView(View):
 #------------------------------------------------------------------------------------------
 #setting up the user
 #------------------------------------------------------------------------------------------
-class ExerciseConfigurationClassView(View):
-	template_name= 'users/exercise_configuration/table_lesson.html'
+class UserConfigurationClassView(View):
+	template_name= 'users/user_configuration/table_lesson.html'
 
 	def post(self, request, *args, **kwargs):
 		exercise_det=Exercise_det.objects.get(id = self.kwargs['pk'])
@@ -246,17 +260,52 @@ class ExerciseConfigurationClassView(View):
 				   }
 		return render(request, self.template_name, context)
 
+class UserConfigurationSawLessonView(View):
+
+	def get(self, request, *args, **kwargs):
+		try:
+			lesson = Lesson_det.objects.get(id=self.kwargs['id_lesson'])
+		except Lesson_det.DoesNotExist:
+			messages.success(request, 'La clase que desea manipular fue eliminada o no existe', extra_tags='alert-danger')
+			return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])
+
+		if lesson.saw == True:
+			messages.success(request, 'La clase ya fue vista', extra_tags='alert-danger')
+			return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])
+		
+		
+		lesson.saw = True
+		lesson.save()
+		update_resumen(lesson)
+
+		#Se crea el historial de las personas
+		history_obj = History_det.objects.create(
+													cant_max = lesson.cant_max,
+													cant_in = lesson.cant_in,
+													quota = lesson.quota,
+													day_lesson = lesson.day_lesson,
+													hour_chance = lesson.hour_chance,
+													hour_lesson = lesson.hour_lesson,
+													hour_end = lesson.hour_end, 
+													id_exercise_fk = lesson.id_exercise_fk
+												)
+
+		for users_in_lesson in lesson.id_user_fk.all():
+			history_obj.id_user_fk.add(users_in_lesson)
+
+		return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])
+		
 class DeleteLessonView(View):
 	def get(self, request, *args, **kwargs):
 		try:
 			lesson = Lesson_det.objects.get(id=self.kwargs['id_lesson'])
 		except Lesson_det.DoesNotExist:
 			messages.success(request, 'La clase que desea manipular fue eliminada o no existe', extra_tags='alert-danger')
-			return redirect('lesson:list_lesson_exercise_action')
+			return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])	
 
 		if lesson.saw == True:
 			messages.success(request, 'La clase ya fue vista', extra_tags='alert-danger')
-			return redirect('lesson:list_lesson_exercise_action')
+			return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])	
 			
 		users_in_lesson = lesson.id_user_fk.all() #i get all users that are in the lesson		
 
@@ -271,10 +320,10 @@ class DeleteLessonView(View):
 
 		lesson.delete()
 
-		return redirect('content_user:exercise_configuration_class', pk=self.kwargs['id_exercise_det'])	
+		return redirect('content_user:user_configuration_class', pk=self.kwargs['id_exercise_det'])	
 
-class ExerciseConfigurationPlanView(View):
-	template_name = 'users/exercise_configuration/resumen.html'
+class UserConfigurationPlanView(View):
+	template_name = 'users/user_configuration/resumen.html'
 	context = {}
 
 	def post(self, request, *args, **kwargs):
@@ -282,25 +331,32 @@ class ExerciseConfigurationPlanView(View):
 		form = ConfigurationUserExerciseForm(request.POST, instance=exercise_det)
 		user_to_modific = CustomUser.objects.get(exercise_det__id = self.kwargs['pk'])
 
-		#this "if" is part of a validation
+		#this "if" is part of a validation of reset condition of the user
 		if exercise_det.reset == True:
 			if form.is_valid():
 				form.save()
 				# print("Este es la data" + str(form.cleaned_data))
 				# print("ES VALIDO!")
-				return redirect('content_user:exercise_configuration_class', pk=self.kwargs['pk'])
+				return redirect('content_user:user_configuration_class', pk=self.kwargs['pk'])
+			else:
+				print("no es válido")
+				self.context = {
+							'user_to_modific': user_to_modific,
+							'exercise_det': exercise_det,
+							'form':form
+					   }
+			return render(request, self.template_name, self.context)
 
+		elif exercise_det.reset == False: #esto es por si intenta modificar un usuario que ya ha fue previamente colocado en modo "No reiniciar"
+			messages.success(self.request, 'No se han guardado cambios porque previamente ha activado la opción de "No reiniciar"', extra_tags='alert-danger')
+			print("no es válido")
+			self.context = {
+						'user_to_modific': user_to_modific,
+						'exercise_det': exercise_det,
+						'form':form
+				   }
 
-
-		messages.success(self.request, 'No se han guardado cambios porque previamente ha activado la opción de "No reiniciar"', extra_tags='alert-danger')
-		print("no es válido")
-		self.context = {
-					'user_to_modific': user_to_modific,
-					'exercise_det': exercise_det,
-					'form':form
-			   }
-
-		return render(request, self.template_name, self.context)
+			return render(request, self.template_name, self.context)
 
 
 	def get(self, request, *args, **kwargs):
@@ -315,7 +371,7 @@ class ExerciseConfigurationPlanView(View):
 				   }
 		return render(request, self.template_name, self.context)
 
-class ExerciseConfigurationChangePlanView(View):
+class UserConfigurationChangePlanView(View):
 
 	def post(self, request, *args, **kwargs):
 		
@@ -329,11 +385,11 @@ class ExerciseConfigurationChangePlanView(View):
 			exercise_det.resetter()
 
 			# print("ES VALIDO!")
-			return redirect('content_user:exercise_configuration_plan', pk=self.kwargs['pk'])
+			return redirect('content_user:user_configuration_plan', pk=self.kwargs['pk'])
 		else:
 			print("no es válido")
 			messages.success(self.request, 'El usuario todavía está en una clase de {}. No se puede reinicar el plan de un ejercicio con el usuario en la clase.'.format(exercise_det.name), extra_tags='alert-danger')
-		return redirect('content_user:exercise_configuration_change_plan', pk=self.kwargs['pk'])
+		return redirect('content_user:user_configuration_change_plan', pk=self.kwargs['pk'])
 
 
 	def get(self, request, *args, **kwargs):
@@ -344,7 +400,7 @@ class ExerciseConfigurationChangePlanView(View):
 		plan_actual_exercise = Plan.objects.filter(id_exercise_fk__name__iexact=exercise_det.name)
 		plan_actual_exercise = plan_actual_exercise.union(plan_ninguno)
 
-		form.fields['id_plan_fk'].queryset = plan_actual_exercise
+		form.fields['id_plan_fk'].queryset = plan_actual_exercise.order_by("name")
 		user_to_modific = CustomUser.objects.get(exercise_det__id = self.kwargs['pk'])
 
 		context = {
@@ -352,10 +408,10 @@ class ExerciseConfigurationChangePlanView(View):
 						'exercise_det': exercise_det,
 						'form':form
 				   }
-		return render(request,'users/exercise_configuration/change_plan.html', context)
+		return render(request,'users/user_configuration/change_plan.html', context)
 
-class ExerciseConfigurationHistoryView(View):
-	template_name = 'users/exercise_configuration/table_history.html'
+class UserConfigurationHistoryView(View):
+	template_name = 'users/user_configuration/table_history.html'
 
 	def post(self, request, *args, **kwargs):
 		exercise_det 	= Exercise_det.objects.get(pk=self.kwargs['pk'])
@@ -407,20 +463,8 @@ class ExerciseConfigurationHistoryView(View):
 				   }
 		return render(request, self.template_name, context)
 
-
-
-class DeleteHistoryView(View):
-	
-	def get(self, request, *args, **kwargs):
-		
-		history_det = History_det.objects.get(pk=self.kwargs['id_history'])
-		history_det.delete()
-
-		return redirect('content_user:exercise_configuration_history', pk=self.kwargs['id_exercise_det'])
-
-
-class ExerciseConfigurationResetView(View):
-	template_name = 'users/exercise_configuration/reset.html'
+class UserConfigurationResetView(View):
+	template_name = 'users/user_configuration/reset.html'
 
 	def post(self, request, *args, **kwargs):
 		exercise_det= Exercise_det.objects.get(id=self.kwargs['pk'])
@@ -431,7 +475,7 @@ class ExerciseConfigurationResetView(View):
 			form.save()
 			# print(form.cleaned_data)
 			# print("ES VALIDO!")
-			return redirect('content_user:exercise_configuration_class', pk=self.kwargs['pk'])
+			return redirect('content_user:user_configuration_class', pk=self.kwargs['pk'])
 		else:
 			print("no es válido")
 			context = {
